Moving calendar to FormExercisePage
